Phase 4a: Extract middleware logic to serv/app/middleware.py
- Extracted MiddlewareManager class (~816 lines) from serv/_app.py
- Moved complete middleware and error handling system
- Includes middleware stack execution, error handler registration
- Template rendering for error responses with content negotiation
- Default 404, 405, and 500 error handlers with dev mode support
- Updated App class to use MiddlewareManager delegation pattern
- Added middleware manager to serv/app module exports
- All 298 tests passing, no functionality changes
- Maintains complete backward compatibility
