+++ serv/_app.py
@@ -1,9 +1,7 @@
 import asyncio
 import contextlib
-import json
 import logging
 import sys
-import traceback
 from asyncio import Task, get_running_loop
 from collections import defaultdict
 from collections.abc import AsyncIterator, Awaitable, Callable
@@ -24,11 +22,11 @@ from asgiref.typing import (
 )
 from bevy import Inject, get_registry, injectable
 from bevy.containers import Container
-from jinja2 import Environment, FileSystemLoader
 
+from serv._routing import Router
+from serv.app.middleware import MiddlewareManager
 from serv.config import load_raw_config
 from serv.database import DatabaseManager
-from serv.exceptions import HTTPMethodNotAllowedException, ServException
 from serv.extensions import Listener
 from serv.extensions.importer import Importer
 from serv.extensions.loader import ExtensionLoader
@@ -36,7 +34,6 @@ from serv.injectors import inject_request_object, inject_websocket_object
 from serv.protocols import AppContextProtocol, EventEmitterProtocol, RouterProtocol
 from serv.requests import Request
 from serv.responses import ResponseBuilder
-from serv._routing import HTTPNotFoundException, Router
 
 logger = logging.getLogger(__name__)
 
@@ -240,10 +237,9 @@ class App(EventEmitterProtocol, AppContextProtocol):
         self._registry = get_registry()
         self._container = self._registry.create_container()
         self._async_exit_stack = contextlib.AsyncExitStack()
-        self._error_handlers: dict[
-            type[Exception], Callable[[Exception], Awaitable[None]]
-        ] = {}
-        self._middleware = []
+
+        # Initialize middleware manager
+        self._middleware_manager = MiddlewareManager(dev_mode=dev_mode)
 
         # Handle backward compatibility for extension_dir parameter
         actual_extension_dir = extension_dir if extension_dir is None else extension_dir
@@ -259,7 +255,6 @@ class App(EventEmitterProtocol, AppContextProtocol):
         self._database_manager = DatabaseManager(self._config, self._container)
 
         self._init_container()
-        self._register_default_error_handlers()
         self._init_extensions(
             self._config.get("extensions", self._config.get("extensions", []))
         )
@@ -292,10 +287,6 @@ class App(EventEmitterProtocol, AppContextProtocol):
         self._container.instances[EventEmitterProtocol] = self._emit
         self._container.instances[AppContextProtocol] = self
 
-    def _register_default_error_handlers(self):
-        self.add_error_handler(HTTPNotFoundException, self._default_404_handler)
-        self.add_error_handler(HTTPMethodNotAllowedException, self._default_405_handler)
-
     @property
     def dev_mode(self) -> bool:
         """Get the current development mode setting."""
@@ -305,6 +296,8 @@ class App(EventEmitterProtocol, AppContextProtocol):
     def dev_mode(self, value: bool) -> None:
         """Set the development mode setting."""
         self._dev_mode = value
+        # Keep middleware manager in sync
+        self._middleware_manager.dev_mode = value
 
     def on_shutdown(self, callback: Callable[[], Awaitable[None]]):
         """Add a callback to be called when the application is shutting down."""
@@ -391,7 +384,7 @@ class App(EventEmitterProtocol, AppContextProtocol):
             app.add_error_handler(Exception, generic_error_handler)
             ```
         """
-        self._error_handlers[error_type] = handler
+        self._middleware_manager.add_error_handler(error_type, handler)
 
     def add_middleware(self, middleware: Callable[[], AsyncIterator[None]]):
         """Add middleware to the application's middleware stack.
@@ -479,7 +472,7 @@ class App(EventEmitterProtocol, AppContextProtocol):
             Middleware is executed in LIFO (Last In, First Out) order during request
             processing, and FIFO (First In, First Out) order during response processing.
         """
-        self._middleware.append(middleware)
+        self._middleware_manager.add_middleware(middleware)
 
     def add_extension(self, extension: Listener):
         if hasattr(extension, "__extension_spec__") and extension.__extension_spec__:
@@ -553,311 +546,6 @@ class App(EventEmitterProtocol, AppContextProtocol):
                     await self._database_manager.shutdown_databases()
                     await self._async_exit_stack.aclose()
 
-    def _get_template_locations(self) -> list[Path]:
-        """Get the template locations for this app.
-
-        Returns a list of paths to search for templates.
-        """
-        return [Path.cwd() / "templates", Path(__file__).parent / "templates"]
-
-    def _render_template(self, template_name: str, context: dict[str, Any]) -> str:
-        """Render a template with the given context.
-
-        Args:
-            template_name: Name of the template to render
-            context: Context to render the template with
-
-        Returns:
-            Rendered template as a string
-        """
-        template_locations = self._get_template_locations()
-        env = Environment(loader=FileSystemLoader(template_locations))
-
-        # Try to load the template
-        try:
-            template = env.get_template(template_name)
-        except Exception:
-            logger.exception(f"Failed to load template {template_name}")
-            # Special case for error templates - provide a fallback
-            if template_name.startswith("error/"):
-                status_code = context.get("status_code", 500)
-                error_title = context.get("error_title", "Error")
-                error_message = context.get("error_message", "An error occurred")
-
-                return f"""
-                <!DOCTYPE html>
-                <html>
-                <head>
-                    <title>{status_code} {error_title}</title>
-                    <style>
-                        body {{ font-family: Arial, sans-serif; line-height: 1.6; padding: 20px; }}
-                        h1 {{ color: #d00; }}
-                        pre {{ background: #f4f4f4; padding: 10px; border-radius: 5px; }}
-                    </style>
-                </head>
-                <body>
-                    <h1>{status_code} {error_title}</h1>
-                    <p>{error_message}</p>
-                </body>
-                </html>
-                """
-            raise
-
-        # Render the template
-        return template.render(**context)
-
-    @injectable
-    async def _default_error_handler(
-        self,
-        error: Exception,
-        response: Inject[ResponseBuilder],
-        request: Inject[Request],
-    ):
-        logger.exception("Unhandled exception", exc_info=error)
-
-        # Check if the error is a ServException subclass and use its status code
-        status_code = (
-            getattr(error, "status_code", 500)
-            if isinstance(error, ServException)
-            else 500
-        )
-        response.set_status(status_code)
-
-        # Check if the client accepts HTML
-        accept_header = request.headers.get("accept", "")
-        if "text/html" in accept_header:
-            # Use HTML response
-            response.content_type("text/html")
-
-            # Enhanced traceback for development mode
-            if self._dev_mode:
-                # Get full traceback with context
-                tb_lines = traceback.format_exception(
-                    type(error), error, error.__traceback__
-                )
-                full_traceback = "".join(tb_lines)
-
-                # Also include exception chain if present
-                if error.__cause__ or error.__context__:
-                    full_traceback += "\n\n--- Exception Chain ---\n"
-                    if error.__cause__:
-                        cause_tb = traceback.format_exception(
-                            type(error.__cause__),
-                            error.__cause__,
-                            error.__cause__.__traceback__,
-                        )
-                        full_traceback += f"Caused by: {''.join(cause_tb)}"
-                    if error.__context__ and error.__context__ != error.__cause__:
-                        context_tb = traceback.format_exception(
-                            type(error.__context__),
-                            error.__context__,
-                            error.__context__.__traceback__,
-                        )
-                        full_traceback += f"During handling of: {''.join(context_tb)}"
-            else:
-                full_traceback = "".join(traceback.format_exception(error))
-
-            context = {
-                "status_code": status_code,
-                "error_title": "Error",
-                "error_message": "An unexpected error occurred.",
-                "error_type": type(error).__name__,
-                "error_str": str(error),
-                "traceback": full_traceback,
-                "request_path": request.path,
-                "request_method": request.method,
-                "show_details": self._dev_mode,
-            }
-
-            html_content = self._render_template("error/500.html", context)
-            response.body(html_content)
-        elif "application/json" in accept_header:
-            # Use JSON response
-            response.content_type("application/json")
-            error_data = {
-                "status_code": status_code,
-                "error": type(error).__name__,
-                "message": str(error)
-                if self._dev_mode
-                else "An unexpected error occurred.",
-                "path": request.path,
-                "method": request.method,
-            }
-
-            if self._dev_mode:
-                # Enhanced traceback for JSON response in dev mode
-                tb_lines = traceback.format_exception(
-                    type(error), error, error.__traceback__
-                )
-                error_data["traceback"] = tb_lines
-
-                # Include exception chain
-                if error.__cause__:
-                    cause_tb = traceback.format_exception(
-                        type(error.__cause__),
-                        error.__cause__,
-                        error.__cause__.__traceback__,
-                    )
-                    error_data["caused_by"] = cause_tb
-                if error.__context__ and error.__context__ != error.__cause__:
-                    context_tb = traceback.format_exception(
-                        type(error.__context__),
-                        error.__context__,
-                        error.__context__.__traceback__,
-                    )
-                    error_data["context"] = context_tb
-
-            response.body(json.dumps(error_data))
-        else:
-            # Use plaintext response
-            response.content_type("text/plain")
-            if self._dev_mode:
-                # Full traceback in plaintext for dev mode
-                tb_lines = traceback.format_exception(
-                    type(error), error, error.__traceback__
-                )
-                full_traceback = "".join(tb_lines)
-                error_message = f"{status_code} Error: {type(error).__name__}: {error}\n\nFull Traceback:\n{full_traceback}"
-            else:
-                error_message = f"{status_code} Error: An unexpected error occurred."
-            response.body(error_message)
-
-    @injectable
-    async def _default_404_handler(
-        self,
-        error: HTTPNotFoundException,
-        response: Inject[ResponseBuilder],
-        request: Inject[Request],
-    ):
-        response.set_status(HTTPNotFoundException.status_code)
-
-        # Check if the client accepts HTML
-        accept_header = request.headers.get("accept", "")
-        if "text/html" in accept_header:
-            # Use HTML response
-            response.content_type("text/html")
-            context = {
-                "status_code": HTTPNotFoundException.status_code,
-                "error_title": "Not Found",
-                "error_message": error.args[0]
-                if error.args
-                else "The requested resource was not found.",
-                "error_type": "NotFound",
-                "request_path": request.path,
-                "request_method": request.method,
-                "show_details": False,
-            }
-
-            html_content = self._render_template("error/404.html", context)
-            response.body(html_content)
-        elif "application/json" in accept_header:
-            # Use JSON response
-            response.content_type("application/json")
-            error_data = {
-                "status_code": HTTPNotFoundException.status_code,
-                "error": "NotFound",
-                "message": "The requested resource was not found.",
-                "path": request.path,
-                "method": request.method,
-            }
-            response.body(json.dumps(error_data))
-        else:
-            # Use plaintext response
-            response.content_type("text/plain")
-            response.body(
-                f"404 Not Found: The requested resource ({request.path}) was not found."
-            )
-
-    @injectable
-    async def _default_405_handler(
-        self,
-        error: HTTPMethodNotAllowedException,
-        response: Inject[ResponseBuilder],
-        request: Inject[Request],
-    ):
-        response.set_status(HTTPMethodNotAllowedException.status_code)
-
-        allowed_methods_str = (
-            ", ".join(error.allowed_methods) if error.allowed_methods else ""
-        )
-        if error.allowed_methods:
-            response.add_header("Allow", allowed_methods_str)
-
-        # Check if the client accepts HTML
-        accept_header = request.headers.get("accept", "")
-        if "text/html" in accept_header:
-            # Use HTML response
-            response.content_type("text/html")
-            context = {
-                "status_code": HTTPMethodNotAllowedException.status_code,
-                "error_title": "Method Not Allowed",
-                "error_message": error.args[0]
-                if error.args
-                else "The method used is not allowed for the requested resource.",
-                "error_type": type(error).__name__,
-                "error_str": str(error),
-                "request_path": request.path,
-                "request_method": request.method,
-                "allowed_methods": allowed_methods_str,
-                "show_details": False,
-            }
-
-            html_content = self._render_template("error/405.html", context)
-            response.body(html_content)
-        elif "application/json" in accept_header:
-            # Use JSON response
-            response.content_type("application/json")
-            error_data = {
-                "status_code": HTTPMethodNotAllowedException.status_code,
-                "error": "MethodNotAllowed",
-                "message": error.args[0]
-                if error.args
-                else "The method used is not allowed for the requested resource.",
-                "path": request.path,
-                "method": request.method,
-                "allowed_methods": error.allowed_methods
-                if error.allowed_methods
-                else [],
-            }
-            response.body(json.dumps(error_data))
-        else:
-            # Use plaintext response
-            response.content_type("text/plain")
-            message = (
-                error.args[0]
-                if error.args
-                else f"The method used is not allowed for the requested resource {request.path}."
-            )
-            response.body(f"405 Method Not Allowed: {message}")
-
-    @injectable
-    async def _run_error_handler(self, error: Exception, container: Inject[Container]):
-        response_builder = container.get(ResponseBuilder)
-        if not response_builder._headers_sent:
-            response_builder.clear()
-
-        handler_key = type(error)
-        handler = self._error_handlers.get(handler_key)
-        if not handler:
-            for err_type, hnd in self._error_handlers.items():
-                if isinstance(error, err_type):
-                    handler = hnd
-                    break
-        handler = handler or self._default_error_handler
-
-        try:
-            await container.call(handler, error)
-        except Exception as e:
-            logger.exception(
-                "Critical error in error handling mechanism itself", exc_info=True
-            )
-            if handler is not self._default_error_handler:
-                e.__context__ = error
-                ultimate_response_builder = container.get(ResponseBuilder)
-                if not ultimate_response_builder._headers_sent:
-                    ultimate_response_builder.clear()
-                await container.call(self._default_error_handler, e)
-
     async def _lifespan_iterator(self, receive: Receive):
         event = {}
         while event.get("type") != "lifespan.shutdown":
@@ -895,15 +583,19 @@ class App(EventEmitterProtocol, AppContextProtocol):
 
                 # Run middleware stack
                 try:
-                    await self._run_middleware_stack(
-                        container=container, request_instance=request
+                    await self._middleware_manager.run_middleware_stack(
+                        container=container,
+                        request_instance=request,
+                        emit_callback=self.emit,
                     )
                 except Exception as e:
                     error_to_propagate = e
 
                 # Handle any errors that occurred
                 if error_to_propagate:
-                    await container.call(self._run_error_handler, error_to_propagate)
+                    await self._middleware_manager.run_error_handler(
+                        error_to_propagate, container=container
+                    )
 
                 await self.emit(
                     "app.request.end", error=error_to_propagate, container=container
@@ -913,7 +605,7 @@ class App(EventEmitterProtocol, AppContextProtocol):
                 logger.exception(
                     "Unhandled exception during request processing", exc_info=e
                 )
-                await container.call(self._run_error_handler, e)
+                await self._middleware_manager.run_error_handler(e, container=container)
                 await self.emit("app.request.end", error=e, container=container)
 
             finally:
@@ -927,110 +619,6 @@ class App(EventEmitterProtocol, AppContextProtocol):
                     logger.error(
                         "Exception during final send_response", exc_info=final_send_exc
                     )
-
-    async def _run_middleware_stack(
-        self, container: Container, request_instance: Request
-    ):
-        stack = []
-        error_to_propagate = None
-        router_instance = container.get(Router)
-
-        for middleware_factory in self._middleware:
-            try:
-                # For middleware functions, use container.call to properly inject dependencies
-                # Don't await the result since it's an async generator
-                middleware_iterator = container.call(middleware_factory)
-                await anext(middleware_iterator)
-            except Exception as e:
-                logger.exception(
-                    f"Error during setup of middleware {getattr(middleware_factory, '__name__', str(middleware_factory))}",
-                    exc_info=True,
-                )
-                error_to_propagate = e
-                break
-            else:
-                stack.append(middleware_iterator)
-
-        if not error_to_propagate:
-            await self.emit(
-                "app.request.before_router",
-                container=container,
-                request=request_instance,
-                router_instance=router_instance,
-            )
-            try:
-                resolved_route_info = router_instance.resolve_route(
-                    request_instance.path, request_instance.method
-                )
-                if not resolved_route_info:
-                    raise HTTPNotFoundException(
-                        f"No route found for {request_instance.method} {request_instance.path}"
-                    )
-
-            except Exception as e:
-                logger.info(
-                    f"Router resolution resulted in exception: {type(e).__name__}: {e}"
-                )
-                error_to_propagate = e
-
-            else:
-                handler_callable, path_params, route_settings = resolved_route_info
-
-                # Create a branch of the container with route settings
-                with container.branch() as route_container:
-                    # Add route settings to the container using RouteSettings
-                    from serv._routing import RouteSettings
-
-                    route_container.add(RouteSettings, RouteSettings(**route_settings))
-
-                    # Ensure essential dependencies are available in route container
-                    # Copy from parent container if missing
-                    for dep_type in [Request, ResponseBuilder, Container, Router]:
-                        try:
-                            parent_instance = container.get(dep_type)
-                            # Try to get from route container to see if it's already there
-                            try:
-                                route_container.get(dep_type)
-                            except:
-                                # Not found in route container, add it
-                                route_container.add(dep_type, parent_instance)
-                        except:
-                            # Parent doesn't have it, skip
-                            pass
-
-                    try:
-                        await route_container.call(handler_callable, **path_params)
-                    except Exception as e:
-                        logger.info(
-                            f"Handler execution resulted in exception: {type(e).__name__}: {e}"
-                        )
-                        error_to_propagate = e
-
-            await self.emit(
-                "app.request.after_router",
-                container=container,
-                request=request_instance,
-                error=error_to_propagate,
-                router_instance=router_instance,
-            )
-
-        for middleware_iterator in reversed(stack):
-            try:
-                if error_to_propagate:
-                    await middleware_iterator.athrow(error_to_propagate)
-                    error_to_propagate = None
-                else:
-                    await anext(middleware_iterator)
-            except StopAsyncIteration:
-                pass
-            except Exception as e:
-                logger.exception("Error during unwinding of middleware", exc_info=True)
-                if error_to_propagate:
-                    e.__context__ = error_to_propagate
-                error_to_propagate = e
-
-        if error_to_propagate:
-            raise error_to_propagate
 
     async def _handle_websocket(self, scope: Scope, receive: Receive, send: Send):
         """Handle WebSocket connections."""
