+++ serv/app/__init__.py
@@ -3,9 +3,11 @@
 # Re-export from existing modules for backward compatibility
 try:
     from serv._app import App
+    from serv.app.middleware import MiddlewareManager
 
     __all__ = [
         "App",
+        "MiddlewareManager",
     ]
 except ImportError:
     # If direct imports fail, will be populated later during refactoring
